requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
sqlalchemy==2.0.31
aiosqlite==0.20.0
asyncpg==0.29.0